rage: rename reporter to stream
Summary:
My original plan was to expand reporter to take on the responsibility of reporting progress, but it turned out Rust borrow checker is not happy about this deign since I need `write!(stream)` to work, which requires a mutable reference while keeping an immutable reference in `StepGuard` I added later.

As a result I have to add the progress reporting struct directly to `WatchmanRage`. This diff just rename `Reporter` to `Stream` to better reflect what it does.

Reviewed By: MichaelCuevas

Differential Revision: D39187638

fbshipit-source-id: 1e7a0d624978f5a5e5c52f435cf15745db8e3721