Fix type scanners for clang 15/16
Summary: LLVM staging for clang 15/16 will sometimes produce a DW_TAG_structure_type for the same Indexer<...> type in a compile_unit without a corresponding DW_TAG_variable. This causes the type scanners to sometimes (indeterministic based on which thread picks the indexer) choose an invalid type structure from the dwarf. This change adds a check if the object is a valid indexer before picking it.

Reviewed By: MatzeB

Differential Revision: D40033710

fbshipit-source-id: aadfc2cfd0e21cdb8e6f5270be4acc048bfe0833