Fix a phi handling bugs in PRE
Summary:
The PRE algorithm initially starts with an underlying state defined by presence
of IncRef/DecRef and other instructions (marking blocks with local availability
and anticipability).

Then it monotonically propagates the available and anticipated bits throughout
the CFG, which includes monotonically setting the phiPropagate bits, until
convergence is achieved.

However, when we reach the pre_apply() phase, we may realize that it is
impossible to eliminate some IncRefs or DecRefs if they are only partially
redundant and there is no place to insert balancing instructions. When that
happens, we *remove* some bits representing the underlying knowledge.

Instead of running the whole converge algorithm from scratch, we enqueue the
directly affected blocks into the worklist and continue execution of the
algorithm, this time expecting it to converge monotonically backwards until the
next pre_apply() attempt.

Mixing forward and backward monotonic operations is somewhat sketchy, but it
works as long as we do only one direction at a time (which is the case here)
and it assumes that each internal routine is agnostic of the direction. This
was not the case with pre_adjust_for_phis().

Assume the following CFG:

  A   B
   \ /
    C (phi)
   / \
  D   E   G
       \ /
        F

Assume that A and B produce availability (e.g. DecRefs), D and F produce
anticipability (e.g. IncRefs). During the initial convergence, the phi at C
gets marked as phiPropagate, as it gets both availability and anticipability,
letting them propagate in both directions. Once we reach pre_apply(), we are
ready to kill both DecRefs and IncRefs, as long as we can deal with F, which
doesn't get availability from G. If we can't deal with that, we mark F as no
longer providing anticipability. We now run the convergence in the opposite
direction. This causes C to revert to only partial anticipability, which ceases
the reason for phiPropagate bit. But due to the bug, we keep the bit on. This
still turns off the propagation of anticipability towards A and B (this
requires full anticipability in pre_compute_anticipated()), but it doesn't turn
off propagation of availability to D. Hence we remove the IncRef at D, without
taking any other balancing action.

Fix this by making pre_adjust_for_phis() work in both directions.

Reviewed By: binliu19, oulgen

Differential Revision: D40204574

fbshipit-source-id: 8e847b829779e226539669585944c26fdaa4ac21