Do not leak supported refs across block merges
Summary:
When merging two states of `(lower_bound, unsupported_refs)` pairs, take the
minimum of lower_bound and supported_refs, rather than minimum of lower_bound
and maximum of unsupported_refs.

Previously, the merge of (5, 3) and (4, 2) states would result in (4, 3),
reducing the number of stronger supported references from 2 to 1. After this
diff, the result will be (4, 2), properly keeping track of supported refs.

This should improve tracking of lower bounds especially in large CFGs.

Reviewed By: ricklavoie

Differential Revision: D40215347

fbshipit-source-id: 0414e1e5e84d7be3a4c088244dfc0bbf8e22d7b3