Timeout message shouldn't reference connection
Summary:
The existing code references the host and port from the connection, but this is a timeout message so the connection didn't get established.  The fields do appear to be correct but we shouldnt' be assuming this.  Change the code to reference the host and port from the connection key that was requested instead.  This is the correct thing to do here.

Note that this _might_ resolve T134264467

Differential Revision: D40230330

fbshipit-source-id: 8db204eaf2441d0b1d228dbebed72c7b5b67240f