Utilities to switch executors to ThrottledLifoSem
Summary:
Add static methods in `{CPU,EDF}ThreadPoolExecutor` to create queue/semaphore implementations with `LifoSem` and `ThrottledLifoSem`, to avoid some boilerplate to conditionally support both when constructing an executor.

This is mostly to allow to use the ternary operator, as `make_unique<...>` would require additional casts to yield compatible types.

Reviewed By: Gownta, luciang

Differential Revision: D40144624

fbshipit-source-id: b081cc8ed81090a04f0e4792262271a3fc9bee83