Ban refinements in is/as expr via NAST check
Summary:
HackC already fatals when a type refinements appears (recursively) inside
an is/as expression, so do the same in Hack.  We don't need types to tell that,
so do this as part of Naming AST checks (NAST) using a `iter_with_state` visitor
that can recursively pass down the information
whether we are (deep) inside a hint that represent the RHS of an is/as-expression.

The same approach can be reused for banning other constructs for which TypeStructure isn't implemented in the runtime, so make the name of the check general.

Reviewed By: hgoldstein

Differential Revision: D40239918

fbshipit-source-id: d7aa557cc7d9570ab7bb7658ce447e3860066a80