Ban refinements in RHS of type const via NAST check
Summary:
HackC already fatals when a type refinements appears (recursively) inside
an RHS of a type const, so do the same in Hack.  Since we don't need types to tell that,
do this as part of Naming AST checks (NAST) using a `iter_with_state` visitor,
which can recursively pass down the information
whether we are (deep) inside a hint.

Reviewed By: mpu

Differential Revision: D40253129

fbshipit-source-id: 7e3dde630b66b2b0cc6acdcf74135b26f78471f8