Ban refinements in RHS of type alias via NAST check
Summary:
HackC already fatals when a type refinements appears (recursively) inside
an RHS of a type alias, so do the same in Hack. E.g.
- `type TBad1 = Box with { ctx C = [] }`
- `type TBad2 = shape('nested' => Box with { type T = int })`

Since we don't need types to tell that,
do this as part of Naming AST checks (NAST) using a `iter_with_state` visitor,
which can recursively pass down the information
whether we are (deep) inside a hint.

Reviewed By: mpu

Differential Revision: D40253861

fbshipit-source-id: dddf08d2edff1d79f2bc1c91200609425536fec1