Allow requireResourcePools to relax ordering restrictions on server setup
Summary:
There is currently a restriction that if setupThreadManager is called before setInterface then the proper runtime resource pool checks cannot be done and so resource pools is run-time disabled out of caution. This can make it hard to refactor code which needs the threading resources available when it is creating the handler. Where it is easy we have added setters to supply the thread manager instead of providing it in ctor but there are other cases where the dependency is more complex.

This change allows us to migrate via code like
```
if (useResourcePoolFlagsSet()) {
  server->requireResourcePools();
}

... general setup

server->setupThreadManager(); // In resource pools mode this will now be allowed to continue even though we can't do runtime checks yet
auto tm = server->getThreadManager_deprecated(); // In the resource pools mode this will be a thread manager wrapper around default executor

... handler setup which requires tm

server->setInterface(handler);

...
```

```server->setup()``` eventually gets called. In the resource pools case the runtime checks will be re-run and can fail with a fatal error but that will only happen if we have modified the servers code to call requireResourcePools and we should only do that for services which we know (have tested) will pass the run-time checks. The service could be subsequently modified in such a way that it fails the run time checks - eg an interaction added - but that should be caught by rolling out that change.

Reviewed By: sazonovkirill

Differential Revision: D38620268

fbshipit-source-id: 1474e970b8fcc2c528f02ad5d05ec8acfcaae467