Update SSLSessionTest to support TLS 1.3 resumption
Summary:
Session tickets are handled differently in TLS 1.3 than TLS 1.2. They aren't available when the client handshake success callback is invoked. They arrive afterwards. We need to attach a readcallback and wait for them to be populated.

For tests to succeed for both TLS 1.2 and TLS 1.3 versions, we can install a `SessionLifecycleCallbacks` instance that will have its `onNewSession` method invoked when the corresponding socket receives session tickets.

Reviewed By: anhuang

Differential Revision: D40266752

fbshipit-source-id: 0eecce4880237e8448d6daaf8a4413a4edb7ab09