Clean up hdf library so it can be built without folly or boost
Summary:
We'd like to be able to build hackc as a pure Rust build using Cargo.  It can build simple C++ files but linking against boost or folly is tricky.

Luckily with a few tweaks the HDF utility code that we use from HHVM really doesn't need either of those.

Reviewed By: edwinsmith

Differential Revision: D39948459

fbshipit-source-id: cdd1adbc1bb172fb2dc8018fdea5bd524d5e2919