Move Type::__EMPTY__ check into switch statement
Summary:
Move `Type::__EMPTY__` check into switch statement.

The problem is that when building with `-Werror, -Wswitch-enum`, the warning will break the build.

It's not obvious to me why we have to explicitly check this outside of the switch statement, so here I just moved them into the switch statement to fix the warning.

Reviewed By: NSProgrammer

Differential Revision: D40286214

fbshipit-source-id: b202f38bd5b68b580ff44eb842b9e1e4e72c1ba0