Fix subtyping with the `this` type
Summary:
This diff is a completeness fix for type refinements.

A while ago the `this` type in a class used to be a
Tdependent locl type. We decided in D25898947 (https://github.com/facebook/hhvm/commit/e4de89e78d5884d05e07dbe5dfc5d769182b8d84) to
instead represent it using a Tgeneric. This reduced
code duplication quite a bit and simplified the
treatment of bounds, but it also had the side effect
of "forgetting" that `this` is special in that it can
only stand for the class type of a runtime object,
much like expression-dependent types.

When implementing type refinements we built early
on support for expression-dependent types but
forgot that `this` ought to behave in a similar manner.

This defect is fixed in the present diff. The code of
Typing_type_member had to change a bit to be
able to deal with multiple bounds (`this` can have
many), but that is for great good anyway. I also
introduced a variant type that reunites `this` and
expression-dependent types; it is only used by type
refinement code currently.

Two incompleteness tests are fixed.

Reviewed By: vsiles

Differential Revision: D40261134

fbshipit-source-id: cb90b180823f37ee695bcc7163677d2b738e311e