Finer processing of constraints involving refinements
Summary:
This diff is a completeness fix for class refinements.

When processing constraints involving refinements,
the strategy since D39097858 (https://github.com/facebook/hhvm/commit/9b64e14564aefae3ed001707a41d5c7eb5d2f55a) has been to leave the
refinements attached to class types. The main
motivation for doing so is that we do not want
constraints of the form `T <: C with { ... }`
to be broken into `T <: has_type_member(...)`
that could not be internalised later on in the type
parameters env.

This strategy is sound, but it has proved too coarse
as some legit programs in D39374601 (https://github.com/facebook/hhvm/commit/8c6ebd195b01317315ba3c35a54f7a7aaf5869aa) could not
be checked.

In the present diff, we refine subtyping by actually
breaking down refinement checks when we know
they can be discharged on the spot and not generate
`has_type_member(...)` constraints types. One
such case is when the LHS of the `as` constraint is
a class type.

Reviewed By: vsiles

Differential Revision: D40297741

fbshipit-source-id: 558ca9596883cf16555999e3063636c2099fb146