Mark HH\Asio\backtrace return as nullable
Summary:
`HH\Asio\backtrace` can return `Array()` if the awaitable is already finished:
https://www.internalfb.com/code/fbsource/[4ce6ce571557e21e99cfa1ba7eb6b6d06429a03d]/fbcode/hphp/runtime/ext/asio/ext_asio.cpp?lines=148-150

Turns out that `Array()` in C++ translates to `null` in PHP, so the stub/HHI for the method is currently incorrect and needs updating. This shouldn't result in any runtime changes since the method was already able to return `null`, and this more accurately reflects the actual typing.

Reviewed By: jano

Differential Revision: D39703357

fbshipit-source-id: 18c427b12384128945f6e783ef58dcbd94d4c567