add internal_error when we can't find a prop
Summary:
We've noticed a constant level of typecheck exceptions originating from `nastInitCheck.ml` for a while. It spiked during chenmela's oncall in July, and has recently spiked again: see this [workplace post](https://fb.workplace.com/groups/hacklang.oncall/permalink/517295090237047/).

The exception's being thrown in the function `class_prop_pos`, which we call when we're trying to find the `pos` of class properties that aren't initialized. In `class_prop_pos`, this exception is thrown when we've found the class from which the prop originates, but we're unable to find the `c_var` in that class's AAST that matches the prop.

Since we know it's an issue, this diff changes the error from throwing an exception to registering an internal error. Hopefully this will help us debug the error.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D40278676

fbshipit-source-id: b193383fefc60129886a506e87f5a9b316e40823