Handle nullable class type
Summary:
Previously, for the case "$foo?->bar()" where "Foo::bar()" is a memoized instance method, HAGVAH prints it out as "?Foo->bar" (in column "global_variables" of XDB "error_details").

As requested by Basic Ads, the question mark shall be omitted to be consistent with runtime checker (i.e. CIPP Events). In addition, as HAGVAH is not interested in the case where $foo is null, it makes sense to remove the prefix "?" from the printed class type.

Reviewed By: yuxuanchen1997

Differential Revision: D40309816

fbshipit-source-id: cc24ef9511ad6ebd2225fd17ed1011745ac1bbf0