Make hhi library work with cargo build
Summary:
The rust hhi library provides the ability to write out the HHI files to a path.  Unfortunately it's actually a bit complex to gather the HHI files (particularly the HSL ones - the HHVM ones are pretty easy).

This code is only used by HackC for building a decl provider - so in a non-fbcode build just provide no HHIs for now.

Differential Revision: D39948462

fbshipit-source-id: 054185cee241d9353e3ddec4fb60f78958487d71