Rework handling of interfaces/traits/abstract classes in HHBBC
Summary:
HHBBC's handling of interfaces/traits/abstract classes wasn't entirely
correct. The biggest issue is that when calculating a class' subclass
list, it would (usually) leave off any such classes. The assumption
was that since they're not instantiable, they can be ignored. This is
*sometimes* true, but not in other situations. For example. you can
call static methods on a non instantiable class. By simply ignoring
them, you leave out possible method resolutions. For a long time we
got away with this, but as analysis has gotten better (in particular,
with the interface handling improvements), it's become possible to
write trivial programs which crash HHBBC or produce incorrect results.

First introduce some terminology. A class is called "regular" if it's
not an interface/trait/abstract class/enum class. This is a normal,
instantiable (at least in principal) class. A "non-regular" class is
an interface/trait/abstract class/enum class. These cannot be
instantiated, though they can have static methods called on them. Also
classes can inherit methods from them, which can be called.

The key observation is that whether we want to ignore non-regular
classes or not is dependent on context. In fact, it boils down to
whether the type we're doing an operation on is an "Obj" or a
"Cls". If the type is an "Obj", we've instantiated it, therefore it
must be a regular class. This corresponds, roughly, to HHBBC's
previous logic. If the type is a "Cls", then the class could be
regular or non-regular. In that case, we need to consider all
classes. "Cls" presents another wrinkle. In some situations, even if
you have a "Cls", you might know that the Cls came from an Obj (for
example by calling clsobj()), and therefore you only need to deal with
regular classes. This isn't just an optimization. It's needed to
preserve monotonicity. So, to handle this, we add an extra bit of
state to DCls to indicate whether it represents just regular classes,
or all classes.

This bit of state is propagated to all the relevant couldB/subtypeOf
operations in the type-system. Whether a given class is a subtype of
another could vary depending on which subset of subclasses we're
considering. So, add the necessary logic to the comparison operations
(and add a few more which are required). This also affects the
canonicalization logic (as a simple example, Obj=Trait should
canonicalization to Bottom, since you cannot instantiate a
trait). There are many corner cases here.

Subclass list in ClassInfo now contains *all* subclasses. It's the
responsibility of those iterating it to skip non-regular classes or
not (depending on context). Traits are somewhat a special case. The
"subclasses" of a trait are the classes which implement the trait (if
the trait wasn't flattened). So, traits need to be special cases when
iterating the subclass list. Sometimes you want to iterate it,
sometimes you want to treat it as empty.

Besides the type-system, the other major complication is for method
resolution. Again, whether you want to consider all subclasses or just
the regular subclasses is context dependent. Logically, we'd like to
create two FuncFamilies for each method, one for each case. However,
this naive scheme is too expensive memory-wise. First, to save memory,
we only create one FuncFamily for each method, as before. The
"possible funcs" list is *all* methods. The user of the FuncFamily is
responsible for iterating the list and skipping entries as
appropriate. Since the return type and static info for the FuncFamily
can be different depending on the subset of methods considered, we
allocate space in the FuncFamily for both results. We detect
situations where the results cannot differ (this is common) and avoid
allocating extra memory in that case. In most cases, we can store a
single entry in the methodFamilies map for both the "all subclasses"
case and the "regular subclasses" case. In fact, they can only differ
for non-regular base classes. In that case, we use a (smaller)
auxiliary map.

The method/ctor resolution logic is updated to use the new method
family data. The logic is tightened up a bit, and made to consider
corner cases, like having an "exact" non-regular class that we're
trying to look up a non-static method on. The input type to the
resolution logic is now relevant, as whether it's an Obj or a Cls will
determine how the lookup is done. "Special" method families have
special inheritance logic and sometimes dispatch to magic runtime
functions which aren't in the bytecode. For that reason we have to
treat those more conservatively. We're now able to definitively assert
in some cases that a method won't resolve and therefore attempting to
call it will produce a Bottom. This (along with the improved
canonicalization) required a bit more handling of Bottom in the interp
logic.

In order to make the method family logic work properly for interfaces,
add interface methods to the interface ClassInfo's method table at the
same time as the rest of classes. Up until now, they were added after
the fact when building func families. No good reason for this.

Switch the method family tables over to use sorted_vector_map instead
of hphp_fast_map. We have a *lot* of method entries and the extra
space consumed by the hphp_fast_maps was considerable.

Add a new type of Attr, called AttrNoOverrideRegular. This has a
similar meaning to AttrNoOverride, except it's weaker in that it means
the class is not overridden by a *regular* class. The runtime can use
this instead of AttrNoOverride if it knows it has an instance of the
class.

Add a bunch of new gtests/unit tests to test type-system or exercise
various problematic situations I found while developing all this.

Reviewed By: jano

Differential Revision: D40253770

fbshipit-source-id: 84f39abf157274cf21dfe024863aae28c6e2ab39