native facts support
Summary:
This diff adds Native Facts support for the `require class` trait constraint.

The API is unchanged, apart for `require class` constraints being explicitly stored in `$type['requireClass']` and being used to compute `subtypes` and `supertypes`.

In several API functions (eg. `supertypes` of a trait), it is not possible to distinguish if a class is required by a trait via `require extends` or `require class` (as in HHVM `getRequirementNames` reflection function - note: it would be preferable to keep the two APIs in sync).

Reviewed By: jthemphill

Differential Revision: D40144845

fbshipit-source-id: 000ef3d6f6572f39f2046783c04088f771560dd7