Add warning and optional error when subscription names aren't unique
Summary:
Watchman's subscription management logic implicitly assumes subscription names are unique within a single connection, but this is not enforced - which leads to bugs and surprising behaviour.

Here we start enforcing the above assumption, as a warning by default or as an error if the `enforce_unique_subscription_names` config option is set. The error will become the default behaviour in the future.

Reviewed By: chadaustin

Differential Revision: D40207028

fbshipit-source-id: b71e7c9a88a3bdfd1c9aafb7baacd96fcbc90469