Go to definition for class refinements
Summary:
When asking "Go to Definition" on the type
constant name in a refinement, the IDE will
now jump to the class' type constant.

It is possible that, in the future, we allow
refining types that are not known to have
the refined type constant (sub-classes
may define it and satisfy the refinement).
In this case, Go to Definition shouldn't
move. In the current implementation,
we should always have somewhere to
jump to (see D38941765 (https://github.com/facebook/hhvm/commit/14a26df4cd363f0ed48183e4552707030a865c50)).

Reviewed By: strub-meta

Differential Revision: D40340200

fbshipit-source-id: 892d44df02924a7a3d5d06d3df38e5ae865673b3