Fix scratch register used on the function interception code path
Summary:
It was using rarg(0), and all the php_return_regs() are live at that point. This
is a problem on ARM because rarg(0) is also rret_data(). So replace rarg(0) with
rarg(3), which works for both x64 and ARM, and add assertions to  catch
such bugs in the future.

Reviewed By: jano

Differential Revision: D40329054

fbshipit-source-id: a8a9e690625420eca35f9415d758ed0e115736d9