RFC: Expose systemlib decls in full repo builds
Summary:
The previous diffs in this stack exposed systemlib decls to:
* Incremental file compilation
* Compiling systemlib itself at HHVM startup time

This diff indexes all of the decls in `s_builtin_symbols` akin to how files in the repo itself are indexed. They're indexed with an empty file name, which I did specifically to avoid attempts to parse / emit bytecode for said files as part of `Package::resolveOnDemand`: we'll unconditionally parse / emit the bytecode for systemlib.

Reviewed By: edwinsmith

Differential Revision: D39754321

fbshipit-source-id: 3a4fa9d30823293ae9b5570040bc03b0ffec6ecf