Improve exception details for `Future_failure`
Summary:
The `error` union inside a `Future_failure` exception uses a first-class module within the `External` case. As a result, OCaml does not render the inner error (it is existentially bound) and we don't get any useful information about these exceptions.

This diff registers a pretty printer for the `Future_failure` exception which uses `error_to_string` to render the inner `Error.t`

Reviewed By: hgoldstein

Differential Revision: D40224847

fbshipit-source-id: 825cbe48c1005b11bed6489583ce932151789419