only catch Unix_error in connect_to_monitor instead of all exceptions
Summary:
When the client is stuck in the connect to monitor loop, Ctrl+C does not work because the handler raises an exception which is caught in that loop. To fix that, we only catch the exceptions actually thrown by the loop, which according to production logs are only Unix_error exception (whereas the exception thrown on Ctrl+C is an Exit_status).

I've also included nits and doc block fixing.

Reviewed By: nt591

Differential Revision: D40345037

fbshipit-source-id: 1e9f5557dc9c0a59331cd8269b1a50f177a6fb4c