heap_vector_types: improve compatibility with maps whose reference isn't `const value_type&`
Summary: I have an in-progress fancy map that works like the upcoming std::flat_map, where reference is actually a pair-of-references type. Converting to const value_type& here was forcing a copy of the map entry. templating each method works the same in the typical case without pessimizing the fancy stuff.

Reviewed By: Gownta

Differential Revision: D40209621

fbshipit-source-id: aa4e1d62c1e2e6e10ba8c58246b6ae581de39f13