sorted_vector_types: improve compatibility with maps whose reference isn't `const value_type&`
Summary: I have an in-progress fancy map that works like the upcoming `std::flat_map`, where `reference` is actually a pair-of-references type. Converting to `const value_type&` here was forcing a copy of the map entry. `const auto&` works the same for normal code while not causing problems with fancy stuff either.

Reviewed By: Gownta

Differential Revision: D40209620

fbshipit-source-id: 3b0350f5ef2fab1469225c45dfdc73bd7865977a