Fix some private method analysis in HHBBC
Summary:
The changes to deal with abstract classes properly inadvertantly broke
some of the analysis around private methods. I thought it was actually
making things safer, as we were now including private methods in func
families (we weren't before).

However, this missed out on some subtletly around how private method
lookup is done. If the current context class defines a private method,
and you call that method from an instance of the context class, it
will always call that private method, even if it the method was
overridden in a child class.

This means that a private method on a non-regular class, even if
overridden in every regular sub-class, can still be called. Therefore
it is *not* correct to exclude it from the regular subset list in func
familes. Fix the incorrect analysis by always treating private methods
as if coming from regular classes. This meant we had to loosen some
invariants we were asserting on.

This change is sufficient for correctness. However, we can improve the
analysis of private method calls in situations where we were
pessimizing before. When resolving a method, it's relatively
straight-forward to check the above condition and infer the exact
private method that will be called. We just need to check the current
context, see if the current class derives from the context, then do a
method lookup on the context.

With this, we no longer need the pessimization around the
"hasPrivateAncestor" bool on methods.

Reviewed By: oulgen

Differential Revision: D40400527

fbshipit-source-id: f8147482fafeb8eae709024e0385685eea938c6b