Read JustKnob load_state_natively_v4 value to override hh.conf value only for certain Sandcastle jobs
Summary:
More context here: https://fb.workplace.com/groups/hacklang.eng/permalink/9581962815185581/

We are auditing all usages of fallbacks to full inits and want to either 1)
replace the full init approach with saved state or 2) explicitly allow full
inits by adding the --no-load flag. This diff will roll out option 1) to certain
Sandcastle jobs controlled by the settings in the JustKnobs UI.

When the JustKnob is true, hh_server will have the ability to fetch saved
state in the event of hh_server failing in the middle of its normal operations
for the SC job. This behavior is already enabled for all devservers and ODs.
For safety, I will first only roll out to one Sandcastle job, www-static-analysis,
and only a percentage of job instances with this alias.

Reviewed By: nt591

Differential Revision: D40251129

fbshipit-source-id: 5222fd7407209b7e2b683bab23a48dd08e6ef12e