Use output directory as extern-worker working directory by default
Summary:
With trait flattening now using extern-worker, the
performance difference between slow and fast disks is quite large when
using the subprocess impl. By default, extern-worker uses /tmp (or
whatever the system configured temp directory is). This tends to be a
slower disk on a lot of systems.

If an explicit working directory is not specified, use the output
directory instead. Since we're going to write the output there, it
tends to be faster.

Reviewed By: oulgen

Differential Revision: D40459999

fbshipit-source-id: 5f4e52d29217b96c2a289918cc60bd3176b74415