Ensure hh_single_type_check --root does not blow away rust_provider_backend SharedMem config overrides
Summary: Rust_provider_backend requires shm_use_sharded_hashtbl to be enabled, but when `--root` is provided, we use the sharedmem config in .hhconfig (without JustKnobs overrides). This causes us not to enable the `shm_use_sharded_hashtbl` setting. Delay the overriding so that it is applied after `--root` constructs the config.

Reviewed By: shayne-fletcher

Differential Revision: D40482867

fbshipit-source-id: ce379d06a7eb8c6ca935222f2135b425dcfcc796