add folly::is_sorted_vector_set/map helper traits
Summary:
sorted_vector_set/map needs to be carefully used, because insertion/deletion/update to those types could cause `O(n)` resorting, which is very expensive.

Sometimes we want to check if a set/map type is based on sorted_vector, and if so we change the strategy to update those types, for example batch the inserts into a bulk_insert() call.

Adding folly::is_sorted_vector_set/map helper traits so we can use them to test if a set/map type is based on sorted_vector.

Reviewed By: Gownta

Differential Revision: D40374727

fbshipit-source-id: d163023869a7617fed78cc5503be2d692c0b4f93