Automatically infer glob constraints for whole-of-repo queries (`all`, fresh instance `since`)
Summary:
# Problem

When the Eden watcher handles a `since` query in a fresh Watchman instance (or an `all` query), we currently send an expensive `globFiles` request to Eden that globs the entire repo (or the entirety of `relative_root`, if available). This is known to degrade Watchman and Eden's stability on developer machines, to the extent that one unaware client can render the entire environment unusable.

# Solution

Here, we teach Watchman to generate narrower glob requests to Eden automatically, by looking at the query expression. To do this, we introduce the concept of computing a *glob upper bound* for the query, and use this upper bound as the argument to `globFiles`.

As a result, we can answer queries more efficiently without requiring client-side changes - at least, for queries that already contain a term expressing a prefix match that we understand. Other clients will need to rewrite their queries to benefit from this change, but this approach still has the advantage of not complicating the query language (vs something like supporting multiple `relative_root`s).

# Details

The new behaviour is gated by the `eden_enable_glob_upper_bounds` config, which currently defaults to false.

For a given expression node, a glob upper bound is a set of glob patterns that is guaranteed to match every path that the node will match (and possibly other paths too). If no bounding is possible, we don't return an upper bound. (Note that this differs from returning an empty set as the upper bound.)

Our heuristics for translating expression nodes to glob upper bounds are very basic and can be improved (see comments), but I *think* we can try to ship the current implementation as a first version. Generally, we bias towards bounding the entire query with the fewest glob patterns, and with patterns that represent prefix matches (as opposed to bounding the depth and/or suffix of a path).

Reviewed By: xavierd

Differential Revision: D39689958

fbshipit-source-id: 1ca6a27c2003be9de483c4d4bd398e677c53bbb8