Simplify magic bool conversion tracking in HHBBC
Summary:
We have a complicated infrastructure (making use of pointers to member
functions) to track which classes might contain magic
methods. Unfortunately this is all unused because the only magic
method supported anymore is __toBoolean and that is special cased
anyways. Get rid of find_magic_methods() and related. When we want to
know if a class might contain a magic bool conversion, we can just do
the check directly (it's simple).

Reviewed By: oulgen, jano

Differential Revision: D40540994

fbshipit-source-id: cd27523dc3486e5c14c24196a5be51d0d213edb1