Generate rust_to_ocaml(and) attributes
Summary:
Mutual recursion in type definitions is opt-in in OCaml:

```
type param = ...

type expr =
  | Call of ...
  | Lambda of param list * stmt list

and stmt =
  | Expr of expr
  | Return of expr
```

Without using the `and` keyword to introduce the `stmt` type, the type name `stmt` would be unbound in the definition of `expr`.

Rust does not have such a requirement, so we either need to communicate to `rust_to_ocaml` where mutual recursion exists with an attribute or make `rust_to_ocaml` clever enough to detect it. This stack opts for the former.

This diff has hh_oxidize generate `rust_to_ocaml(and)` attributes wherever the `and` keyword is necessary in a mutually recursive type definition.

Reviewed By: edwinsmith

Differential Revision: D40495120

fbshipit-source-id: 557485ebcd3678ccd45c754a2a77b1bfe1e36ca2