ir::TypedValue
Summary: Convert use of hhbc::TypedValue in the IR to (a newly added) ir::TypedValue. This lets us use UnitStringId to represent the strings within the TypedValue instead of arena-based allocation. We can also track Dict and Vec as IndexMap and Vec which is easier to manipulate than raw Slice.

Reviewed By: edwinsmith

Differential Revision: D40490647

fbshipit-source-id: 0b8c6b8a9129588c08865df51790cf3806e741e3