Improve IR adata support
Summary: In HHBC the adata is per-unit.  This is good for caching and sharing but makes manipulation of functions more difficult. Switch to storing the data directly in the Constant so they can be manipulated without having to pass around a mutable Unit (but as an Arc so they can still be shared when known).

Reviewed By: edwinsmith

Differential Revision: D40488914

fbshipit-source-id: e3036fad45ec3409e863bbdd95bcc51e99b61f7d