Update rollout flags to more accurately reflect which server events actually got an override for the load_state_natively value in hh.conf
Summary:
Currently, the actual value of the JustKnob controlling whether to override the hh.conf value gets dropped between the time it's checked and the time rollout_flags are logged. This means we could be logging load_state_natively_v4=true in the rollout_flags field for cases that are not included in the rollout (devservers/ondemands) as well as cases that are included in the rollout (www-static-analysis, to start, and later, any switch value that is set to True). Then, we can't do any meaningful analysis on scuba data by filtering, in the usual way, on "rollout_flags contains load_state_natively_v4=true".

This is kind of unfortunate because it means this diff, which creates a separate "override_load_state_natively" value to capture the value we're interested in, just adds more confusion to a diff that was already a bit complicated (D40251129 (https://github.com/facebook/hhvm/commit/693fc7af9b5ebbc601b9acc634fad9dd4709ef2c)); however, I think it's the most straightforward given that diff's implementation and I've tried to be really clear in my diff summaries and code comments. I'd love to know how this could be improved though!

Reviewed By: nt591

Differential Revision: D40540209

fbshipit-source-id: 2f07cb56ac572aabcb745f05044a6ec56f844a47