Increase the limit on the number of keys of a StructDict
Summary:
This diff lifts the limit on the maximum number of keys in a StructDict with UnalignedTVLayout. StructDicts with number of keys <= 255 still have the exact same size, ie the iterator positions of these StructDicts are still 1 byte. Bigger StructDicts have 2 byte iterator positions.
The number of colors required to color all fields of StructDicts uniquely is at least as much as the the max number of fields in a StructDict. The increase in field size makes the coloring table size too large and sparse and causes memory and CPU regression. We set the max number of color lower than the max number of fields and only color a prefix of the fields uniquely. If a color table miss occurs for a non-unique color, the field map in the layout needs to be checked. The optimal number of uniquely colored fields, given a max number of color and a set of layouts, is determined by a binary search.

Reviewed By: ricklavoie

Differential Revision: D37969333

fbshipit-source-id: 0d6a51886ea773ad02d9710766d6865d332e7a73