Add `void` types to all declarations of `Iterator::next()` and `Iterator::rewind()`
Summary:
I've been trying a build of D40371013 on WWW. Most of the Hack errors can be dealt with, but it surfaces a few issues in Hack's `.hhi` files which can't be `HH_FIXME`'d away:

```
Typing[4341] The method rewind is not compatible with the overridden method [1]
-> Expected void [2]
-> But got dynamic [1]

/tmp/hh_server/hhi_2a562b63/stdlib/builtins_spl.hhi:218:19
    216 |   public function next() {}
[1] 217 |   public function recoverFromCorruption() {}
    229 |   public function rewind() {}

/tmp/hh_server/hhi_2a562b63/interfaces.hhi:145:29
    143 |    * This rewinds back to the first element of the `Iterator`.
    144 |    */
[2] 145 |   public function rewind(): void;
    146 |   /**
    147 |    * Checks to see if the current iterator position is valid.
```

I think the issue is that `dynamic` can't be unified with `void`, or there is no `~void` type, or something. There is probably a principled way to fix this in the implementation of Sound Dynamic. Regardless, I like clean `hh` output, I like easy fixes that are still correct, and this seems like an easy fix that leads to clean `hh` output and is still correct.

Reviewed By: vassilmladenov

Differential Revision: D40537399

fbshipit-source-id: 84a028e56e1f067bcefd4abb296c84eee629d2dd