Add test case for false negatives of global reads
Summary:
Currently, global reads are reported only when a global object (either a static object or a memoized function that returns an object) is directly used on the RHS of some assignment, e.g.
- `$a = Foo::$obj_prop`
- `$a = memoized_func()`

However, there are many global reads that are not covered yet, which will be handled in the next diff. Here this diff creates a new test case that shows the current false negatives:
- a global variable that is not a reference to some object (e.g. a primitive) is used, e.g. `$a = Foo::$int_prop` or `$a = (Foo::$obj_prop)->int_prop` or `$a = (memoized_func())->int_prop`
- a global variable is used in some operation, e.g. `$a = (Foo::$prop)->int_prop + 1`
- a global variable is used in the condition of if statement or loops, e.g. `if (Foo::$int_prop > 0) { ... }`
- a global variable is used in the parameter of some function call, e.g. `func_call(Foo::$prop)`
- a global variable is returned from a function call, e.g. `return (Foo::$prop)->int_prop + 1`.

Reviewed By: lavenzg

Differential Revision: D40568029

fbshipit-source-id: 7547b58dac6866095c3319e92c791f8c71e4713b