Recognize simple singleton patterns for global writes
Summary:
As suggested by our customer (MI), there are some common patterns for the detected global writes, which are benign in most cases. Thus, this diff introduces the following patterns:
  - Singleton (i.e. write to a global variable whose value is null)
  - CounterIncrement (i.e. increase a global variable by 1 or -1)
  - WriteEmptyOrNUll (i.e. assign null or empty string to a global variable)
  - WriteLiteral (i.e. assign a literal to a global variable)
  - WriteGlobalToGlobal (i.e. assign a global variable to another one)

Currently, all the patterns above are related with global writes, and later some patterns for global reads may be introduced as well.

Specifically, this diff starts with a simple case of the singleton pattern, in which a global object (or its mutable property) is written under the condition that its value is null.

To do:
- Handle a more complex case of the singleton pattern, in which the global object is a collection (e.g. dictionary) and functions like `idx` and `C\contains_key` are used.
- To handle other patterns (e.g. WriteLiteral), we will need to track the reference of non-global variables as well. To be more precise, we need a general pointer analyzer, in which the data source of every local variable is maintained (not just the local variables that point to global variables as we do now).

Reviewed By: minjang

Differential Revision: D40181325

fbshipit-source-id: 475ce79d14df7fc2e6df38873a29388394e7d000